test(bigquery-table-desc) added unit test for desc string cleaning
