+++ metadata-ingestion/src/datahub/ingestion/source/bigquery_v2/bigquery.py
@@ -40,6 +40,9 @@ from datahub.ingestion.source.bigquery_v2.bigquery_audit import (
     BigQueryTableRef,
 )
 from datahub.ingestion.source.bigquery_v2.bigquery_config import BigQueryV2Config
+from datahub.ingestion.source.bigquery_v2.bigquery_helper import (
+    unquote_and_decode_unicode_escape_seq,
+)
 from datahub.ingestion.source.bigquery_v2.bigquery_report import BigQueryV2Report
 from datahub.ingestion.source.bigquery_v2.bigquery_schema import (
     BigqueryColumn,
@@ -1073,7 +1076,7 @@ class BigqueryV2Source(StatefulIngestionSourceBase, TestableSource):
 
         dataset_properties = DatasetProperties(
             name=datahub_dataset_name.get_table_display_name(),
-            description=self.unquote_and_decode_unicode_escape_seq(table.comment)
+            description=unquote_and_decode_unicode_escape_seq(table.comment)
             if table.comment
             else "",
             qualifiedName=str(datahub_dataset_name),
@@ -1383,21 +1386,3 @@ class BigqueryV2Source(StatefulIngestionSourceBase, TestableSource):
             self.config.start_time,
             self.config.end_time,
         )
-
-    def unquote_and_decode_unicode_escape_seq(
-        self,
-        string: str,
-        leading_quote: str = '"',
-        trailing_quote: Optional[str] = None,
-    ) -> str:
-        """
-        If string starts and ends with a quote, unquote it and decode Unicode escape sequences
-        """
-        trailing_quote = trailing_quote if trailing_quote else leading_quote
-
-        if string.startswith(leading_quote) and string.endswith(trailing_quote):
-            string = string[1:-1]
-
-        cleaned_string = string.encode().decode("unicode-escape")
-
-        return cleaned_string
